Fix DB Schema

+++ src/main/java/eu/europa/ec/dgc/gateway/entity/FederatedEntity.java
@@ -45,6 +45,6 @@ public abstract class FederatedEntity {
     String domain;
 
     @Column(name = "version", columnDefinition = "int default 1", nullable = false)
-    private Long version = 1L;
+    private Integer version = 1;
 
 }
